# job-service-contract/src/main/resources/com/hpe/caf/services/job/swagger.yaml
#
# Copyright 2016-2020 Micro Focus or one of its affiliates.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

swagger: "2.0"
info:
  title: "Job Service"
  version: "1.0"
  description: |
    Allows background operations to be sent to the workers, and for the
    operations to be tracked and controlled.

    You can check the progress of operations, and request to cancel, pause, or resume them.  *NB: Pause and Resume are future
    requirements.*

    The primary use of this functionality is batch document
    processing.  The service can send work to the Batch Worker, and
    then track the progress of that work.
basePath: /job-service/v1
tags:
  - name: Jobs
    description: Job Control
consumes:
  - application/json
produces:
  - application/json

parameters:
  partitionId:
    name: partitionId
    in: path
    type: string
    required: true
    description: Only allow access to jobs in the container with this identifier
    pattern: '^[^.,:;*?!|()]{1,40}$'
  jobId:
    name: jobId
    in: path
    type: string
    required: true
    description: The identifier of the job
    pattern: '^[^.,:;*?!|()]{1,48}$'

paths:
  /partitions/{partitionId}/jobs:
    parameters:
      - $ref: '#/parameters/partitionId'
      - name: CAF-Correlation-Id
        in: header
        type: string
        description: |
          An identifier that correlates events 
          across different services.
      - name: jobIdStartsWith
        in: query
        type: string
        required: false
        description: Only those results whose job id starts with this value will be returned
      - name: statusType
        in: query
        type: string
        required: false
        description: All - no status filter is applied (Default); NotCompleted - only those results with statuses other than Completed will be returned; Completed - only those results with Completed status will be returned; Inactive - only those results with inactive statuses (i.e. Completed, Failed, Cancelled) will be returned; NotFinished - only those results with unfinished statuses (ie. Active, Paused, Waiting) will be returned.
        enum:
          - All
          - NotCompleted
          - Completed
          - Inactive
          - NotFinished
      - name: limit
        in: query
        type: integer
        format: int32
        required: false
        description: The maximum results to return (i.e. page size)
      - name: offset
        in: query
        type: integer
        format: int32
        required: false
        description: The starting position from which to return results (useful for paging)
      - name: sort
        in: query
        type: string
        required: false
        description: 'How to order the returned results, in the format <field>:<direction>.  Allowed values for field: jobId, createTime.  Allowed values for direction: asc, desc.'
        default: createTime:desc
      - name: labelExist
        in: query
        type: string
        required: false
        description: 'Filter to jobs with the specified labels, in the format label=<labelName>,<labelName>'
      - name: filter
        in: query
        type: string
        required: false
        description: 'Filter jobs with the specified criteria, in the format `labels.<label-key> == <label-value>` or `id == 1`'
      - name: legacyDateFormat
        in: query
        type: boolean
        required: false
        description: 'If False the service will return any dates in a human readable date format, i.e. similar to dd/mm/yyyy HH:MM:SS:SSS`'
    get:
      tags:
        - Jobs
      summary: Gets the list of jobs
      description: Returns the list of job definitions defined in the system.
      operationId: getJobs
      responses:
        200:
          description: Returns the list of jobs.
          schema:
            type: array
            items:
              $ref: "#/definitions/job"
  /partitions/{partitionId}/jobs/{jobId}:
    parameters:
      - $ref: '#/parameters/partitionId'
      - name: CAF-Correlation-Id
        in: header
        type: string
        description: |
          An identifier that correlates events 
          across different services.
      - $ref: '#/parameters/jobId'
      - name: legacyDateFormat
        in: query
        type: boolean
        required: false
        description: 'If False the service will return any dates in a human readable date format, i.e. similar to dd/mm/yyyy HH:MM:SS:SSS`'
    get:
      tags:
        - Jobs
      summary: Gets the specified job.
      description: Retrieves information about the specified job.
      operationId: getJob
      responses:
        200:
          description: Returns the job data.
          schema:
            $ref: "#/definitions/job"
        400:
          description: The `jobId` parameter contains invalid characters.
        404:
          description: The specified job is not found.
    put:
      tags:
        - Jobs
      summary: Adds a new job.
      description: |
        Creates the specified job using the job definition included in the http
        body.
      operationId: createOrUpdateJob
      parameters:
        - name: newJob
          description: The definition of the job to create
          in: body
          required: true
          schema:
            $ref: "#/definitions/newJob"
      responses:
        201:
          description: Indicates that the job was successfully created.
        204:
          description: Indicates that the job was successfully updated.
        400:
          description: The `jobId` parameter contains invalid characters.
    delete:
      tags:
        - Jobs
      summary: Deletes the specified job.
      description: Deletes the specified job from the system.
      operationId: deleteJob
      responses:
        204:
          description: Indicates that the job was successfully deleted.
        400:
          description: The `jobId` parameter contains invalid characters.
        404:
          description: The specified job is not found.
  /partitions/{partitionId}/jobs/{jobId}/isActive:
    parameters:
      - $ref: '#/parameters/partitionId'
      - name: CAF-Correlation-Id
        in: header
        type: string
        description: |
          An identifier that correlates events 
          across different services.
      - $ref: '#/parameters/jobId'
    get:
      tags:
        - Jobs
      summary: Checks whether the job is active.
      description: Checks whether the specified job is active.
      operationId: getJobActive
      responses:
        200:
          description: Returns whether the job is active.
          schema:
            type: boolean
          headers:
            CacheableJobStatus:
                description: A response that includes this header with the value ‘true’ is eligible for caching in a cache specific to job status responses.
                type: boolean
        400:
          description: The `jobId` parameter contains invalid characters.
  /partitions/{partitionId}/jobs/{jobId}/cancel:
    parameters:
      - $ref: '#/parameters/partitionId'
      - name: CAF-Correlation-Id
        in: header
        type: string
        description: |
          An identifier that correlates events 
          across different CAF services
      - $ref: '#/parameters/jobId'
    post:
      tags:
        - Jobs
      summary: Cancels the job.
      description: Cancels the specified job.
      operationId: cancelJob
      responses:
        204:
          description: The cancellation has been accepted.
        400:
          description: The `jobId` parameter contains invalid characters.
        404:
          description: The specified job is not found.
  /partitions/{partitionId}/jobStats/count:
    parameters:
      - $ref: '#/parameters/partitionId'
      - name: CAF-Correlation-Id
        in: header
        type: string
        description: |
          An identifier that correlates events
          across different services.
      - name: jobIdStartsWith
        in: query
        type: string
        required: false
        description: Only those results whose job id starts with this value will be returned
      - name: statusType
        in: query
        type: string
        required: false
        description: All - no status filter is applied (Default); NotCompleted - only those results with statuses other than Completed will be returned; Completed - only those results with Completed status will be returned; Inactive - only those results with inactive statuses (i.e. Completed, Failed, Cancelled) will be returned; NotFinished - only those results with unfinished statuses (ie. Active, Paused, Waiting) will be returned.
        enum:
          - All
          - NotCompleted
          - Completed
          - Inactive
          - NotFinished
      - name: filter
        in: query
        type: string
        required: false
        description: 'Filter jobs with the specified criteria, in the format `labels.<label-key> == <label-value>` or `id == 1`'
    get:
      tags:
        - Jobs
      summary: Gets the count of jobs
      description: Returns the count of job definitions defined in the system.
      operationId: getJobsCount
      responses:
        200:
          description: Returns the count of jobs.
          schema:
            type: integer
            format: int64
definitions:
  newJob:
    type: object
    description: Exactly one of task and job should be specified.
    properties:
      name:
        type: string
        description: The name of the job
      description:
        type: string
        description: The description of the job
      externalData:
        # TODO: Should we make this more structured than just string
        #       Name/Value Pairs perhaps
        #       Or possibly even more structured than that - storing in xml
        #       would mean xpath searches could be done later
        type: string
        description: |
          DEPRECATED - Replaced by labels. External data can be associated with the job for use by other
          components.
      task:
        $ref: "#/definitions/worker-action"
      type:
        type: string
        description: Execute the job using a specific job type configured on the service.
      parameters:
        type: object
        description: Input to the selected job type.
      prerequisiteJobIds:
        type: array
        items:
          type: string
        description: List of job identifiers that must be complete prior to the start of this job.
      delay:
        type: integer
        format: int32
        default: 0		
        description: The time in seconds after the prerequisite job identifiers have completed before this job is eligible for running.
      labels:
        type: object
        additionalProperties:
          type: string
        description: |
          A map of meta data associated with this job. Note that the keys can only include alphanumeric, '_', '-' and
          ':' characters
        example:
          "tag:4": "4"
          owner: "bob"
  worker-action:
    type: object
    required:
      - taskClassifier
      - taskApiVersion
      - taskData
      - taskPipe
      - targetPipe
    properties:
      taskClassifier:
        type: string
      taskApiVersion:
        type: integer
        format: int32
      taskData:
        type: object
      taskDataEncoding:
        type: string
        enum:
          - utf8
          - base64
      taskPipe:
        type: string
      targetPipe:
        type: string
  job:
    type: object
    properties:
      id:
        type: string
        description: The job identifier
      name:
        type: string
        description: The name of the job
      description:
        type: string
        description: The description of the job
      externalData:
        type: string
        description: |
          DEPRECATED - Replaced by labels External data can be associated with the job for use by other
          components.
      createTime:
        type: string
        format: date-time
        description: The time the job was created
      lastUpdateTime:
        type: string
        format: date-time
        description: The time the job status or progress last changed
      status:
        type: string
        description: The status of the job
        enum:
          - Active
          - Cancelled
          - Completed
          - Failed
          - Paused
          - Waiting
      percentageComplete:
        type: number
        format: float
        description: Gives an indication of the progress of the job.
      failures:
        type: array
        items:
          $ref: "#/definitions/failure"
        description: Job failure details
      labels:
        type: object
        additionalProperties:
          type: string
        description: |
          A map of meta data associated with this job. Note that the keys can only include alphanumeric, '_', '-' and
          ':' characters
  failure:
    type: object
    properties:
      failureId:
        type: string
      failureTime:
        type: string
        format: date-time
      failureSource:
        type: string
      failureMessage:
        type: string